Call standard math functions with explicit casts

+++ modules/audio_processing/transient/transient_suppressor.cc
@@ -139,9 +139,11 @@ int TransientSuppressor::Initialize(int sample_rate_hz,
   for (size_t i = 0; i < complex_analysis_length_; ++i) {
     mean_factor_[i] =
         kFactorHeight /
-            (1.f + exp(kLowSlope * static_cast<int>(i - kMinVoiceBin))) +
+            (1.f + std::exp(static_cast<double>(
+                       kLowSlope * static_cast<int>(i - kMinVoiceBin)))) +
         kFactorHeight /
-            (1.f + exp(kHighSlope * static_cast<int>(kMaxVoiceBin - i)));
+            (1.f + std::exp(static_cast<double>(
+                       kHighSlope * static_cast<int>(kMaxVoiceBin - i))));
   }
   detector_smoothed_ = 0.f;
   keypress_counter_ = 0;
@@ -352,7 +354,8 @@ void TransientSuppressor::UpdateBuffers(float* data) {
 // If a restoration takes place, the |magnitudes_| are updated to the new value.
 void TransientSuppressor::HardRestoration(float* spectral_mean) {
   const float detector_result =
-      1.f - pow(1.f - detector_smoothed_, using_reference_ ? 200.f : 50.f);
+      1.f - std::pow(static_cast<double>(1.f - detector_smoothed_),
+                     static_cast<double>(using_reference_ ? 200.f : 50.f));
   // To restore, we get the peaks in the spectrum. If higher than the previous
   // spectral mean we adjust them.
   for (size_t i = 0; i < complex_analysis_length_; ++i) {
